requirements.txt
django
djangorestframework
drf-nested-routers
drf-jwt
